+++ CHANGELOG.md
@@ -538,3 +538,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Make whatmod API public
   * `WhatMod.findModName`
   * `WhatMod.findModID`
+* Add outline color to ModColorDisplay